Refactor get-all-the-things logic from SearchBar to utils
